refactor: remove Basic Pitch transcription
