+++ package.json
@@ -25,9 +25,7 @@
     "@radix-ui/react-slider": "^1.3.6",
     "@remix-run/fetch-router": "^0.20.1",
     "@remix-run/route-pattern": "^0.23.0",
-    "@spotify/basic-pitch": "^1.0.1",
     "@tanstack/react-query": "^5.90.16",
-    "@tensorflow/tfjs": "^3.21.0",
     "@tonejs/midi": "^2.0.28",
     "cmdk": "^1.1.1",
     "jszip": "^3.10.1",
